Reorganize testing tasks
